+++ product_code/admission_service/impl/src/main/scala/de/upb/cs/uc4/admission/impl/AdmissionSerializerRegistry.scala
@@ -3,7 +3,7 @@ package de.upb.cs.uc4.admission.impl
 import com.lightbend.lagom.scaladsl.playjson.JsonSerializer
 import de.upb.cs.uc4.admission.impl.actor.AdmissionsWrapper
 import de.upb.cs.uc4.admission.model.{ CourseAdmission, DropAdmission }
-import de.upb.cs.uc4.shared.client.{ SignedProposal, SignedTransaction, UnsignedProposal, UnsignedTransaction }
+import de.upb.cs.uc4.hyperledger.api.model.{ SignedProposal, SignedTransaction, UnsignedProposal, UnsignedTransaction }
 import de.upb.cs.uc4.shared.server.SharedSerializerRegistry
 
 import scala.collection.immutable.Seq