+++ product_code/examreg_service/impl/src/test/scala/de/upb/cs/uc4/examreg/impl/ExamregServiceSpec.scala
@@ -1,7 +1,5 @@
 package de.upb.cs.uc4.examreg.impl
 
-import java.nio.file.Path
-
 import akka.Done
 import akka.actor.ActorSystem
 import akka.stream.Materializer
@@ -11,8 +9,8 @@ import de.upb.cs.uc4.examreg.DefaultTestExamRegs
 import de.upb.cs.uc4.examreg.api.ExamregService
 import de.upb.cs.uc4.examreg.impl.actor.ExamregHyperledgerBehaviour
 import de.upb.cs.uc4.examreg.model.{ ExaminationRegulation, Module }
+import de.upb.cs.uc4.hyperledger.api.model.JsonHyperledgerVersion
 import de.upb.cs.uc4.hyperledger.connections.traits.ConnectionExaminationRegulationTrait
-import de.upb.cs.uc4.shared.client.JsonHyperledgerVersion
 import de.upb.cs.uc4.shared.client.JsonUtility.{ FromJsonUtil, ToJsonUtil }
 import de.upb.cs.uc4.shared.client.exceptions.{ ErrorType, UC4Exception }
 import de.upb.cs.uc4.shared.server.UC4SpecUtils
@@ -23,6 +21,7 @@ import org.scalatest.matchers.should.Matchers
 import org.scalatest.time.{ Seconds, Span }
 import org.scalatest.wordspec.AsyncWordSpec
 
+import java.nio.file.Path
 import scala.concurrent.{ Await, Future }
 
 class ExamregServiceSpec extends AsyncWordSpec
@@ -33,7 +32,7 @@ class ExamregServiceSpec extends AsyncWordSpec
       .withJdbc()
   ) { ctx =>
       new ExamregApplication(ctx) with LocalServiceLocator {
-        override def createActorFactory: ExamregHyperledgerBehaviour = new ExamregHyperledgerBehaviour(config) {
+        override def createHyperledgerActor: ExamregHyperledgerBehaviour = new ExamregHyperledgerBehaviour(config) {
 
           override val walletPath: Path = retrieveFolderPathWithCreation("uc4.hyperledger.walletPath", "/hyperledger_assets/wallet/")
           override val networkDescriptionPath: Path = retrievePath("uc4.hyperledger.networkConfig", "/hyperledger_assets/connection_profile_kubernetes_local.yaml")