+++ product_code/group_service/impl/src/main/scala/de/upb/cs/uc4/group/impl/actor/GroupBehaviour.scala
@@ -4,13 +4,13 @@ import akka.cluster.sharding.typed.scaladsl.EntityTypeKey
 import akka.pattern.StatusReply
 import com.typesafe.config.Config
 import de.upb.cs.uc4.group.impl.commands.AddToGroup
-import de.upb.cs.uc4.hyperledger.commands.{ HyperledgerBaseCommand, HyperledgerCommand, HyperledgerReadCommand, HyperledgerWriteCommand }
 import de.upb.cs.uc4.hyperledger.connections.cases.ConnectionGroup
 import de.upb.cs.uc4.hyperledger.connections.traits.ConnectionGroupTrait
-import de.upb.cs.uc4.hyperledger.{ HyperledgerActorObject, HyperledgerDefaultActorFactory }
+import de.upb.cs.uc4.hyperledger.impl.commands.{ HyperledgerBaseCommand, HyperledgerCommand, HyperledgerReadCommand, HyperledgerWriteCommand }
+import de.upb.cs.uc4.hyperledger.impl.{ HyperledgerActor, HyperledgerActorObject }
 import de.upb.cs.uc4.shared.server.messages.Accepted
 
-class GroupBehaviour(val config: Config) extends HyperledgerDefaultActorFactory[ConnectionGroupTrait] {
+class GroupBehaviour(val config: Config) extends HyperledgerActor[ConnectionGroupTrait] {
 
   /** Creates the connection to the chaincode */
   override protected def createConnection: ConnectionGroupTrait =