+++ product_code/shared/client/src/main/scala/de/upb/cs/uc4/shared/client/exceptions/ErrorType.scala
@@ -20,7 +20,8 @@ object ErrorType extends Enumeration {
   InternalServer, UndeserializableException, //500
   NotImplemented, //501
   HLInternal, //In an InformativeError
-  HLUnknownTransactionId, HLInsufficientApprovals, HLNotFound, HLConflict, HLUnprocessableLedgerState, HLParameterNumberError, //In a GenericError
+  HLUnknownTransactionId, HLInsufficientApprovals, HLNotFound, HLConflict, HLUnprocessableLedgerState, //In a GenericError
+  HLParameterNumberError, HLApprovalDenied, HLRejectionDenied, //In a GenericError
   HLUnprocessableEntity, HLInvalidEntity, //In a DetailedError
   HLInvalidTransactionCall //In a TransactionError
   = Value