+++ product_code/report_service/CHANGELOG.md
@@ -1,3 +1,15 @@
+# [v0.17.2](https://github.com/upb-uc4/University-Credits-4.0/compare/report-v0.17.1...report-v0.17.2) (2021-02-11)
+## Feature
+## Refactor
+## Bugfix
+ - Fixed endpoint name to conform to the API definition
+
+# [v0.17.1](https://github.com/upb-uc4/University-Credits-4.0/compare/report-v0.16.1...report-v0.17.1) (2021-02-08)
+## Feature
+ - Added endpoint for fetching a pdf with the certificate of enrollment
+## Refactor
+## Bugfix
+
 # [v0.16.1](https://github.com/upb-uc4/University-Credits-4.0/compare/report-v0.16.0...report-v0.16.1) (2021-01-28)
 ## Feature
 ## Refactor